tests/wpt/web-platform-tests/mixed-content/imageset.https.sub.html

<!DOCTYPE html>
<html>
  <head>
    <title>Mixed-Content: imageset tests</title>
    <meta charset='utf-8'>
    <meta name="description" content="Test behavior of imageset resources, making sure that resources are not downloaded.">
    <link rel="author" title="Yoav Weiss" href="yoav@yoav.ws">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/preload/resources/preload_helper.js"></script>
    <script>
      var t = async_test('Makes sure imageset blockable resources are not downloaded');
    </script>
  </head>
  <body>
  <img src="http://{{domains[]}}:{{ports[http][0]}}/images/smiley.png?img_src">
  <picture>
    <source srcset="http://{{domains[]}}:{{ports[http][0]}}/images/smiley.png?picture">
    <img>
  </picture>
  <img srcset="http://{{domains[]}}:{{ports[http][0]}}/images/smiley.png?img_srcset">
  <script>
    window.addEventListener("load", t.step_func(function() {
        verifyNumberOfResourceTimingEntries("http://{{domains[]}}:{{ports[http][0]}}/images/smiley.png?img_src", 1);
        verifyNumberOfResourceTimingEntries("http://{{domains[]}}:{{ports[http][0]}}/images/smiley.png?picture", 0);
        verifyNumberOfResourceTimingEntries("http://{{domains[]}}:{{ports[http][0]}}/images/smiley.png?img_srcset", 0);
        t.done();
    }));
  </script>
  </body>
</html>
